+++ tests/testthat/test-calculate_tabulates.R
@@ -22,4 +22,3 @@ test_that("porcentaje por fila", {
   tab <- calculate_tabulates(base = toybase_individual_2016_03,x = 'REGION', y = 'CH04', weights = 'PONDERA', add.percentage ='row')
   expect_equal(tab[1,2], '50.4%')
 })
-